move crate::cli into crate::sudo
because it's sudo specific

also move crate::env into crate::sudo for the same reason

move some data structures from crate::sudo::cli into
sudo::common::context to avoid making crate::sudo::cli effectively
pub(crate)
