+++ src/lib.rs
@@ -1,10 +1,8 @@
 #[macro_use]
 mod macros;
-pub(crate) mod cli;
 pub(crate) mod common;
 pub(crate) mod cutils;
 pub(crate) mod defaults;
-pub(crate) mod env;
 pub(crate) mod exec;
 pub(crate) mod log;
 pub(crate) mod pam;
